feat:完善前后端细节，加入上传者属性，用户可删除自己的文件

+++ internal/apiServer/objects/store.go
@@ -9,7 +9,7 @@ import (
 	"storage/internal/pkg/utils"
 )
 
-func storeObject(r io.Reader, hash string, size int64) (int, error) {
+func storeObject(r io.Reader, hash string, size int) (int, error) {
 	log.Println("api.objects.storeObject")
 	if locate.Exist(hash) { //如果该对象已经存在，直接返回
 		log.Println("api.objects.storeObject1")
